# Use Span.Slice instead of creating new arrays