--- a/appveyor.yml
+++ b/appveyor.yml
@@ -2,6 +2,8 @@ build_script:
   - cmd: dotnet build -c Release --version-suffix r%APPVEYOR_BUILD_NUMBER%
   - cmd: dotnet test plist-cil.test\plist-cil.test.csproj
   - cmd: dotnet run --project plist-cil.benchmark/plist-cil.benchmark.csproj -c Release
+  - cmd: git checkout master
+  - cmd: dotnet run --project plist-cil.benchmark/plist-cil.benchmark.csproj -c Release
 
 on_success:
   - cmd: dotnet pack plist-cil\plist-cil.csproj -c Release --version-suffix r%APPVEYOR_BUILD_NUMBER%