A lot of code in the `BinaryPropertyListParser` works with sections of arrays. Instead of creating a new array and copying that data to the new array, we can now use `Span.Slice`,  which reduces a lot of allocations and improves performance.